Rename : common 앱 이름 수정 #5
- common 앱 이름을 commons로 수정
