+++ config/django/base.py
@@ -16,7 +16,7 @@ ALLOWED_HOSTS = ["*"]
 #                     설치된 앱, 사용하는 앱 config                         #
 # ==================================================================== #
 LOCAL_APPS = [
-    "mung_manager.common.apps.CommonConfig",
+    "mung_manager.commons.apps.CommonConfig",
     "mung_manager.schemas.apps.SchemasConfig",
     "mung_manager.authentications.apps.AuthenticationConfig",
 ]
