+++ mung_manager/authentications/models.py
@@ -7,7 +7,7 @@ from django.contrib.auth.models import (
 from django.db import models
 
 from mung_manager.authentications.managers import UserManager
-from mung_manager.common.base.models import SimpleModel, TimeStampedModel
+from mung_manager.commons.base.models import SimpleModel, TimeStampedModel
 
 
 class User(AbstractBaseUser, TimeStampedModel, PermissionsMixin):
