+++ mung_manager/commons/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CommonConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "mung_manager.common"
+    name = "mung_manager.commons"
