+++ mung_manager/commons/selectors.py
